pyproject.toml
[tool.poetry]
name = "microshop"
version = "0.1.0"
description = ""
authors = ["ribondareva <msbondareva01@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
pydantic = {extras = ["email"], version = "^2.10.3"}
sqlalchemy = {extras = ["asincio"], version = "^2.0.36"}
aiosqlite = "^0.20.0"
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"
websockets = "^14.1"
fastapi-jwt = "^0.3.0"
python-jose = "^3.3.0"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
bcrypt = "^4.2.1"
python-multipart = "^0.0.20"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
